requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
boto3==1.35.88
botocore==1.35.88
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
jmespath==1.0.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.1
python-multipart==0.0.12
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.40.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.32.0
